+++ compact_test.go
@@ -464,6 +464,7 @@ func (erringBReader) MaxTime() int64                       { return 0 }
 type nopChunkWriter struct{}
 
 func (nopChunkWriter) WriteChunks(chunks ...chunks.Meta) error { return nil }
+func (nopChunkWriter) Size() int64                             { return 0 }
 func (nopChunkWriter) Close() error                            { return nil }
 
 func TestCompaction_populateBlock(t *testing.T) {